check authorization at Reply api

+++ src/main/kotlin/com/teamsparta/todo/replies/ReplyServiceImpl.kt
@@ -5,6 +5,7 @@ import com.teamsparta.todo.replies.dtos.DeleteReplyArguments
 import com.teamsparta.todo.replies.dtos.ReplyDto
 import com.teamsparta.todo.replies.dtos.UpdateReplyArguments
 import com.teamsparta.todo.todocards.TodoCardRepository
+import com.teamsparta.todo.users.User
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 
@@ -13,14 +14,13 @@ class ReplyServiceImpl(
     val replyRepository: ReplyRepository,
     val todoCardRepository: TodoCardRepository,
 ) : ReplyService {
-    override fun createReply(createReplyArguments: CreateReplyArguments): ReplyDto {
+    override fun createReply(createReplyArguments: CreateReplyArguments, user: User): ReplyDto {
         val targetTodoCard = todoCardRepository.findByIdOrNull(createReplyArguments.todoCardId)
             ?: throw Exception("target todo card is not found")
 
         val reply = Reply(
             content = createReplyArguments.content,
-            authorName = createReplyArguments.authorName,
-            password = createReplyArguments.password,
+            author = user,
             todoCard = targetTodoCard,
         )
 
@@ -29,12 +29,12 @@ class ReplyServiceImpl(
         return ReplyDto.from(result)
     }
 
-    override fun updateReply(updateReplyArguments: UpdateReplyArguments): ReplyDto {
+    override fun updateReply(updateReplyArguments: UpdateReplyArguments, user: User): ReplyDto {
         val foundReply = updateReplyArguments.id?.let {
             replyRepository.findByIdOrNull(it)
         } ?: throw Exception("target reply is not found")
 
-        foundReply.checkAuthentication(updateReplyArguments.authorName, updateReplyArguments.password)
+        foundReply.checkAuthorization(user)
 
         foundReply.changeContent(updateReplyArguments.content)
 
@@ -43,12 +43,12 @@ class ReplyServiceImpl(
         return ReplyDto.from(foundReply)
     }
 
-    override fun deleteReply(deleteReplyArguments: DeleteReplyArguments) {
+    override fun deleteReply(deleteReplyArguments: DeleteReplyArguments, user: User) {
         val foundReply = deleteReplyArguments.id?.let {
             replyRepository.findByIdOrNull(it)
         } ?: throw Exception("target reply is not found")
 
-        foundReply.checkAuthentication(deleteReplyArguments.authorName, deleteReplyArguments.password)
+        foundReply.checkAuthorization(user)
 
         replyRepository.deleteById(deleteReplyArguments.id)
     }
